Restructure Common.UI.Controls layout and fix Pop animation flash
- Rename TransientSurface.cs to TransientSurface.xaml.cs (templated-control
  code-behind convention, matching KeyVisual).
- Group TransparentWindow.cs and its Showing/Hiding event args under a new
  Window/TransparentWindow folder.
- Move FlyoutWindowHelper from Flyout/ into Window/ (namespace .Flyout -> .Window)
  and update consumers (PowerDisplay, QuickAccess).
- Fix the Pop transition flash: pin the scale center via a composition
  expression animation and align the show timing so the grow completes under
  the fade-in instead of after it.
