+++ src/common/Common.UI.Controls/Window/TransparentWindow/HidingEventArgs.cs
