+++ src/common/Common.UI.Controls/Window/TransparentWindow/ShowingEventArgs.cs
