+++ src/common/Common.UI.Controls/Window/TransparentWindow/TransparentWindow.cs
